% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeSpec.R
\name{common.always.available.stock.genomes}
\alias{common.always.available.stock.genomes}
\alias{#'}
\title{common.always.available.stock.genomes}
\usage{
common.always.available.stock.genomes()
}
\value{
an list of short genome codes
}
\description{
a helper function for mostly internal use, returns the basic sure-to-be-supported
      igv.js stock genomes, the most commonly used, obviating an aws lookup
}
