% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.extractBamContigs}
\alias{.extractBamContigs}
\title{Extract contig information from a BAM file}
\usage{
.extractBamContigs(bamFile)
}
\arguments{
\item{bamFile}{Path to a readable .bam file}
}
\value{
List with contigs (named integer vector) and assembly (string or NULL)
}
\description{
Extract contig information from a BAM file
}
\keyword{internal}
