% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.extractCramContigs}
\alias{.extractCramContigs}
\title{Extract contig information from a CRAM file}
\usage{
.extractCramContigs(cramFile)
}
\arguments{
\item{cramFile}{Path to a readable .cram file}
}
\value{
List with contigs (named integer vector) and assembly (string or NULL)
}
\description{
Extract contig information from a CRAM file
}
\keyword{internal}
