% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.extractVcfContigs}
\alias{.extractVcfContigs}
\title{Extract contig information from a VCF object or file}
\usage{
.extractVcfContigs(vcfDataOrFile)
}
\arguments{
\item{vcfDataOrFile}{A VariantAnnotation VCF object, or a path to a .vcf / .vcf.gz file}
}
\value{
List with contigs (named integer vector) and assembly (string or NULL)
}
\description{
Extract contig information from a VCF object or file
}
\keyword{internal}
