% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.getGenomeSpec}
\alias{.getGenomeSpec}
\title{Retrieve active genome spec for an igvShiny widget instance}
\usage{
.getGenomeSpec(session, id = NULL)
}
\arguments{
\item{session}{Shiny session or NULL}

\item{id}{Character string, the widget element id}
}
\value{
Named list of genome options, or NULL
}
\description{
Retrieve active genome spec for an igvShiny widget instance
}
\keyword{internal}
