% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.getReferenceContigs}
\alias{.getReferenceContigs}
\title{Extract contig names and lengths from genome options}
\usage{
.getReferenceContigs(genomeSpec)
}
\arguments{
\item{genomeSpec}{Named list of genome options, or character string (genome name)}
}
\value{
Named integer vector of contig lengths, or NULL
}
\description{
Extract contig names and lengths from genome options
}
\keyword{internal}
