% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.identifyLikelyAssembly}
\alias{.identifyLikelyAssembly}
\title{Infer the likely assembly of a set of contigs from lengths or metadata}
\usage{
.identifyLikelyAssembly(contigs, assemblyStr = NULL)
}
\arguments{
\item{contigs}{Named integer vector of contig lengths}

\item{assemblyStr}{Optional string from header}
}
\value{
String describing likely assembly or NULL
}
\description{
Infer the likely assembly of a set of contigs from lengths or metadata
}
\keyword{internal}
