% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{.packJunctionAttributes}
\alias{.packJunctionAttributes}
\title{Pack junction attributes into a bed name column}
\usage{
.packJunctionAttributes(tbl)
}
\arguments{
\item{tbl}{A data.frame of junctions.}
}
\value{
A character vector, one packed name field per row.
}
\description{
Pack junction attributes into a bed name column
}
\keyword{igvShiny}
