% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.registerGenomeSpec}
\alias{.registerGenomeSpec}
\title{Register active genome spec for an igvShiny widget instance}
\usage{
.registerGenomeSpec(session, elementId, genomeOptions)
}
\arguments{
\item{session}{Shiny session or NULL}

\item{elementId}{Character string or NULL}

\item{genomeOptions}{Named list of genome options}
}
\description{
Register active genome spec for an igvShiny widget instance
}
\keyword{internal}
