% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{.sanitizeAndMergeOptions}
\alias{.sanitizeAndMergeOptions}
\title{Sanitize and merge track configuration options}
\usage{
.sanitizeAndMergeOptions(baseOptions, userOptions)
}
\arguments{
\item{baseOptions}{A list of default options set by the R function.}

\item{userOptions}{A list of options provided by the user via trackConfig.}
}
\value{
A merged and sanitized list of options ready to be sent to JavaScript.
}
\description{
Sanitize and merge track configuration options
}
\keyword{igvShiny}
