% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCheck.R
\name{.validateTrackReference}
\alias{.validateTrackReference}
\title{Internal validator dispatched by track loaders}
\usage{
.validateTrackReference(
  session,
  id,
  trackName,
  target,
  validateReference = TRUE
)
}
\arguments{
\item{session}{Shiny session or NULL}

\item{id}{Widget element ID}

\item{trackName}{Display name of track}

\item{target}{Track file path or data object}

\item{validateReference}{Logical, whether validation is enabled}
}
\description{
Internal validator dispatched by track loaders
}
\keyword{internal}
