% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeSpec.R
\name{get_basic_genomes}
\alias{get_basic_genomes}
\title{get_basic_genomes}
\usage{
get_basic_genomes()
}
\value{
an list of short genome codes, e.g., "hg38", "dm6"
}
\description{
a helper function for basic genomes, 
obtains the genome codes (e.g. 'hg38')
}
\examples{
bs <- get_basic_genomes()

}
