% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeSpec.R
\name{get_cas_genomes}
\alias{get_cas_genomes}
\title{get_cas_genomes}
\usage{
get_cas_genomes()
}
\value{
an list of short genome codes, e.g., "hg38", "dm6"
}
\description{
a helper function for common always available stock genomes,
obtains the genome codes (e.g. 'hg38')
}
\examples{
cas <- get_cas_genomes()

}
\keyword{utils}
