% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeSpec.R
\name{get_css_genomes}
\alias{get_css_genomes}
\title{get_css_genomes}
\usage{
get_css_genomes(test = FALSE)
}
\arguments{
\item{test}{logical(1) defaults to FALSE}
}
\value{
an list of short genome codes, e.g., "hg38", "dm6", "tair10"
}
\description{
a helper function for mostly internal use, 
obtains the genome codes (e.g. 'hg38') supported by igv.js
}
\examples{
css <- get_css_genomes(test = TRUE)

}
