% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASTrack.R
\name{getUrl,GWASTrack-method}
\alias{getUrl,GWASTrack-method}
\alias{getUrl}
\title{the url of the gwas table}
\usage{
\S4method{getUrl}{GWASTrack}(obj)
}
\arguments{
\item{obj}{An object of class GWASTrack}
}
\value{
character
}
\description{
the url of the gwas table
}
