% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{loadBigWigTrack}
\alias{loadBigWigTrack}
\title{load a bam track which, with index, is served up by http}
\usage{
loadBigWigTrack(
  session,
  id,
  trackName,
  bigwigURL,
  color = "gray",
  trackHeight = 30,
  autoscale,
  min = NA_real_,
  max = NA_real_,
  deleteTracksOfSameName = TRUE,
  displayMode = "EXPANDED"
)
}
\arguments{
\item{session}{an environment or list, provided and managed by shiny}

\item{id}{character string, the html element id of this widget instance}

\item{trackName}{character string}

\item{bigwigURL}{character string http url for the bigwig file, typically very large}

\item{color}{character string, a legal CSS color, or "random", "gray" by default}

\item{trackHeight}{an integer, 30 (pixels) by default}

\item{autoscale}{logical}

\item{min}{numeric, consulted when autoscale is FALSE}

\item{max}{numeric, consulted when autoscale is FALSE}

\item{deleteTracksOfSameName}{logical, default TRUE}

\item{displayMode}{character string, possible values are "EXPANDED" (default),
"SQUISHED" or "COLLAPSED"}
}
\value{
nothing
}
\description{

}
