% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{loadCramTrack}
\alias{loadCramTrack}
\alias{loadCramTrackToHTTP}
\title{load a cram track which, with index, is served up by http}
\usage{
loadCramTrack(
  session,
  id,
  trackName,
  cramURL,
  craiURL,
  deleteTracksOfSameName = TRUE
)
}
\arguments{
\item{session}{an environment or list, provided and managed by shiny}

\item{id}{character string, the html element id of this widget instance}

\item{trackName}{character string}

\item{cramURL}{character string http url for the bam file, typically very large}

\item{indexURL}{character string http url for the bam file index, typically small
#' @param deleteTracksOfSameName logical, default TRUE}
}
\value{
nothing
}
\description{

}
