% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{loadMergedBigWigFromURLs}
\alias{loadMergedBigWigFromURLs}
\title{load a Merged track from URLs}
\usage{
loadMergedBigWigFromURLs(
  session,
  id,
  mergedName,
  trackName1,
  trackName2,
  url1,
  url2,
  color = "gray",
  trackHeight = 30,
  autoscale = TRUE,
  min = 0,
  max = 1,
  autoscaleGroup = -1,
  deleteTracksOfSameName = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{session}{an environment or list, provided and managed by shiny}

\item{id}{character string, the html element id of this widget instance}

\item{mergedName}{character string}

\item{trackName1}{character string}

\item{trackName2}{character string}

\item{url1}{character}

\item{url2}{character}

\item{color}{character string, a legal CSS color, or "random", 
"gray" by default}

\item{trackHeight}{an integer, 30 (pixels) by default}

\item{autoscale}{logical}

\item{min}{numeric, consulted when autoscale is FALSE}

\item{max}{numeric, consulted when autoscale is FALSE}

\item{autoscaleGroup}{numeric(1) defaults to -1}

\item{deleteTracksOfSameName}{logical(1) defaults to TRUE}

\item{quiet}{logical, default TRUE, controls verbosity}
}
\value{
nothing
}
\description{
load a Merged track
}
\examples{
library(igvShiny)
demo_app_file <-
  system.file(package = "igvShiny", "demos", "igvShinyDemo.R")
if (interactive()) {
  shiny::runApp(demo_app_file)
}

}
\keyword{track_loaders}
