% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{loadSession}
\alias{loadSession}
\title{load session json}
\usage{
loadSession(session, id, url)
}
\arguments{
\item{session}{an environment or list, provided and managed by shiny}

\item{id}{character string, the html element id of this widget instance}

\item{url}{character string}
}
\value{
nothing
}
\description{
load session json
}
\examples{
library(igvShiny)
demo_app_file <-
  system.file(package = "igvShiny", "demos", "igvShinyDemo.R")
if (interactive()) {
  shiny::runApp(demo_app_file)
}

}
\keyword{track_loaders}
