% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{redrawIgvWidget}
\alias{redrawIgvWidget}
\title{redraw the igv genome browser element}
\usage{
redrawIgvWidget(session)
}
\arguments{
\item{session}{an environmet or list, provided and managed by shiny}
}
\description{
maybe a relic, unused, originally intended to refresh?
}
