% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{renderIgvShiny}
\alias{renderIgvShiny}
\title{draw the igv genome browser element}
\usage{
renderIgvShiny(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates an HTML widget}

\item{env}{the environment in which to evaluate expr}

\item{quoted}{logical flag indicating if expr a quoted expression}
}
\value{
an output or render function that enables the use of the widget
within Shiny applications
}
\description{
This function is called in the server function of your shiny app
}
\examples{
library(igvShiny)
demo_app_file <-
  system.file(package = "igvShiny", "showcase", "igvShinyDemo.R")
if (interactive()) {
  shiny::runApp(demo_app_file)
}

}
\keyword{igvShiny}
