% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{saveCompressionSession}
\alias{saveCompressionSession}
\title{Save a session}
\usage{
saveCompressionSession(session, id, event_name = "igv_session")
}
\arguments{
\item{session}{an environment or list, provided and managed by shiny}

\item{id}{character string, the html element id of this widget instance}
}
\value{
nothing
}
\description{
Saves a compressed session.
}
\examples{
library(igvShiny)
demo_app_file <-
  system.file(package = "igvShiny", "demos", "igvShinyDemo.R")
if (interactive()) {
  shiny::runApp(demo_app_file)
}

}
\keyword{track_loaders}
