% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASTrack.R
\name{show,GWASTrack-method}
\alias{show,GWASTrack-method}
\title{Display a short summary of a GWASTrack object}
\usage{
\S4method{show}{GWASTrack}(object)
}
\arguments{
\item{object}{An object of class GWASTrack}
}
\value{
nothing; called for the side effect of printing a summary
}
\description{
Display a short summary of a GWASTrack object
}
\examples{
file <-
  system.file(package = "igvShiny", "extdata", "gwas-5k.tsv.gz")
tbl.gwas <- read.table(file, sep = "\t", header = TRUE, quote = "")
track <-
  GWASTrack("gwas 5k", tbl.gwas, chrom.col = 12, pos.col = 13, pval.col = 28)
track

}
\keyword{GWASTrack_class}
