% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_click_and_check}
\alias{test_click_and_check}
\title{test_click_and_check}
\usage{
test_click_and_check(button_id, expected_html_label, app)
}
\arguments{
\item{button_id}{string of button which will be clicked in the test}

\item{expected_html_label}{string which attendance will be expected in html}

\item{app}{AppDriver shinyApp using for test}
}
\description{
a helper function for clicking the button and checking if
second parameter exist in html code
}
