% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASTrack.R
\name{display,GWASTrack-method}
\alias{display,GWASTrack-method}
\alias{display}
\title{display the already constructed and configured track}
\usage{
\S4method{display}{GWASTrack}(obj, session, id, deleteTracksOfSameName = TRUE)
}
\arguments{
\item{obj}{An object of class GWASTrack}

\item{session}{a Shiny session object}

\item{id}{character the identifier of the target igv object in the browser}

\item{deleteTracksOfSameName}{logical to avoid duplications in track names}
}
\value{
nothing
}
\description{
display the already constructed and configured track
}
\examples{
track <-
  GWASTrack(
    "gwas 5k",
    tbl.gwas,
    chrom.col = 12,
    pos.col = 13,
    pval.col = 28
  )
display(track)

}
