% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{removeTracksByName}
\alias{removeTracksByName}
\title{remove tracks from the browser}
\usage{
removeTracksByName(session, id, trackNames)
}
\arguments{
\item{session}{an environment or list, provided and managed by shiny}

\item{id}{character string, the html element id of this widget instance}

\item{trackNames}{a vector of character strings}
}
\value{
nothing
}
\description{
delete tracks on the browser
}
\examples{
library(igvShiny)
demo_app_file <-
  system.file(package = "igvShiny", "demos", "igvShinyDemo.R")
if (interactive()) {
  shiny::runApp(demo_app_file)
}

}
\keyword{igvShiny}
