% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{removeUserAddedTracks}
\alias{removeUserAddedTracks}
\title{remove only those tracks explicitly added by your app}
\usage{
removeUserAddedTracks(session, id)
}
\arguments{
\item{session}{an environment or list, provided and managed by shiny}

\item{id}{character string, the html element id of this widget instance}
}
\value{
nothing
}
\description{
remove only those tracks explicitly added by your app.  stock tracks (i.e.,
Refseq Genes) remain
}
