% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvShiny.R
\name{renderIgvShiny}
\alias{renderIgvShiny}
\title{draw the igv genome browser element}
\usage{
renderIgvShiny(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{not sure...}

\item{env}{not sure...}

\item{quoted}{not sure...}
}
\description{
This function is called in the server function of your shiny app
}
